+++ pyproject.toml
@@ -15,6 +15,7 @@ loguru = "^0.5.3"
 click = "^8.0.1"
 pydantic = {version = "^1.8.2", extras = ["dotenv"]}
 uvicorn = "^0.15.0"
+requests = "^2.26.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.4"
